% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_annotate.R
\name{annotate_polya_data}
\alias{annotate_polya_data}
\title{Annotate poly(A) data frame with ensembl gene and transcript ids from annotation data frame}
\usage{
annotate_polya_data(input_data_frame, annotated_references)
}
\arguments{
\item{input_data_frame}{data frame with poly(A) data}

\item{annotated_references}{data frame with annotated references}
}
\value{
a data.frame with added columns:
\itemize{
\item ensembl_gene_id: ensembl gene id (without version number
\item ensemble_transcript_id: ensembl transcript id (without version number)
\item symbol: gene symbol
}
}
\description{
Annotate poly(A) data frame with ensembl gene and transcript ids from annotation data frame
}
\examples{
\dontrun{

annotate_polya_data(example_valid_polya_table,annotated_references)

}

}
