% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_annotate.R
\name{annotate_references_with_gencode}
\alias{annotate_references_with_gencode}
\title{Annotate references with ensembl gene and transcript ids and gene symbol, parsing gencode headers}
\usage{
annotate_references_with_gencode(input_list)
}
\arguments{
\item{input_list}{list with references element}
}
\value{
a list with 'references' element with added columns:
\itemize{
\item ensembl_gene_id: ensembl gene id (without version number
\item ensemble_transcript_id: ensembl transcript id (without version number)
\item symbol: gene symbol
}
}
\description{
Annotate references with ensembl gene and transcript ids and gene symbol, parsing gencode headers
}
\examples{
\dontrun{

annotate_references_with_gencode(polya_data)

}

}
