% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_annotate.R
\name{annotate_with_annotables}
\alias{annotate_with_annotables}
\title{Annotate polyA predictions using annotables}
\usage{
annotate_with_annotables(polya_data, genome)
}
\arguments{
\item{polya_data}{polya data table to annotate}

\item{genome}{valid genome from annotables to use for annotation}
}
\value{
a \link[tibble]{tibble}
}
\description{
Annotate polyA predictions using annotables
}
