% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_annotate.R
\name{annotate_with_org_packages}
\alias{annotate_with_org_packages}
\title{Title}
\usage{
annotate_with_org_packages(
  polya_data,
  columns_of_annotation = c("GENENAME", "SYMBOL"),
  keytype = "ENSEMBLTRANS",
  organism = "mus_musculus"
)
}
\arguments{
\item{polya_data}{polya data table to annotate}

\item{columns_of_annotation}{which columns to use}

\item{keytype}{whic keytype to use}

\item{organism}{whic organism database to use}
}
\value{
a \link[tibble]{tibble}
}
\description{
Title
}
