% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_helper_functions.R
\docType{data}
\name{axis_elements_size}
\alias{axis_elements_size}
\title{Default theme for ggplot2-based plots in the NanoTail package}
\format{An object of class \code{numeric} of length 1.}
\usage{
axis_elements_size
}
\description{
Default theme for ggplot2-based plots in the NanoTail package
}
\keyword{datasets}
