% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_stats.R
\name{calculate_diff_exp_binom}
\alias{calculate_diff_exp_binom}
\title{Title}
\usage{
calculate_diff_exp_binom(polya_data_summarized, grouping_factor,
  condition1, condition2, alpha = 0.05)
}
\arguments{
\item{alpha}{}
}
\description{
Title
}
