% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_stats.R
\name{calculate_pca}
\alias{calculate_pca}
\title{Calculates PCA using polya predictions or counts}
\usage{
calculate_pca(polya_data_summarized, parameter = "polya_median")
}
\arguments{
\item{polya_data_summarized}{summarized polyA predictions. Generate use \link{summarize_polya}}

\item{parameter}{\itemize{
\item parameter used for PCA calculation. One of: polya_median,polya_mean,polya_gm_mean,counts
}}
}
\value{
pca object
}
\description{
Needs polyA predictions table summarized by \link{summarize_polya} function, using "sample_name" as summary_factors
}
