% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_stats.R
\name{calculate_polya_stats}
\alias{calculate_polya_stats}
\title{Calculate basic statistics}
\usage{
calculate_polya_stats(polya_data, min_reads = 0,
  grouping_factor = "group", condition1 = NA, condition2 = NA)
}
\arguments{
\item{condition2}{}
}
\value{
tibble with statistics
}
\description{
Calculate basic statistics
}
