% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_helper_functions.R
\name{calculate_scaling_vector_for_virutal_gel}
\alias{calculate_scaling_vector_for_virutal_gel}
\title{Calculates scaling vector for virtual gel plotting}
\usage{
calculate_scaling_vector_for_virutal_gel(input_data, groupingFactor)
}
\arguments{
\item{input_data}{input polyA table for calculation of scaling factor (count of reads)}

\item{groupingFactor}{for which factor calculate counts}
}
\value{
named vector
}
\description{
Calculates scaling vector for virtual gel plotting
}
