% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_plots.R
\name{.basic_aesthetics}
\alias{.basic_aesthetics}
\title{Title}
\usage{
.basic_aesthetics(ggplot_object, scale_x_limit_low = NA,
  scale_x_limit_high = NA, scale_y_limit_low = NA,
  scale_y_limit_high = NA, color_palette = "Set1", plot_title = NA,
  color_mode = "color", axis_titles_size = 16)
}
\arguments{
\item{ggplot_object}{ggplot2 object to manipulate asesthetics}

\item{scale_x_limit_low}{lower limit of x continuous scale}

\item{scale_x_limit_high}{upper limit of x continuous scale}

\item{scale_y_limit_low}{lower limit of y continuous scale}

\item{scale_y_limit_high}{upper limit of y continuous scale}

\item{color_palette}{color palette (one from RColorBrewer of ggsci packages)}

\item{plot_title}{Title of the plot}
}
\value{
\link[ggplot2]{ggplot} object
}
\description{
Title
}
