% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_stats.R
\name{.kruskal_polya}
\alias{.kruskal_polya}
\title{Compute Kruskal-Wallis test on single poly(A) data}
\usage{
.kruskal_polya(
  input_data,
  grouping_factor = "sample_name",
  verbose = F,
  verbosity_level = 1
)

.kruskal_polya(
  input_data,
  grouping_factor = "sample_name",
  verbose = F,
  verbosity_level = 1
)
}
\arguments{
\item{input_data}{input data.frame (for single transcript)}

\item{grouping_factor}{which column contains group information}

\item{verbose}{verbose output}

\item{verbosity_level}{how verbose the output should be (levels 1 - little verbosity, or 2 - very verbose)}
}
\value{
data.frame with test statistics

data.frame with test statistics
}
\description{
Compute Kruskal-Wallis test on single poly(A) data

Compute Kruskal-Wallis test on single poly(A) data
}
