% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_stats.R
\name{.polya_stats}
\alias{.polya_stats}
\title{Calculates polyA statistics for single group of reads (for single transcript)}
\usage{
.polya_stats(polya_data, stat_test, grouping_factor, min_reads = 0,
  use_dwell_time = FALSE, custom_glm_formula = NA)
}
\arguments{
\item{polya_data}{\itemize{
\item input data frame with polyA predictions
}}

\item{stat_test}{\itemize{
\item statistical test to use. One of : Wilcoxon, KS (Kolmogorov-Smirnov) or glm (Generalized Linear Model). All tests use log2(polya_length) as a response variable
}}

\item{grouping_factor}{\itemize{
\item factor defining groups (Need to have 2 levels)
}}

\item{min_reads}{\itemize{
\item minimum reads per group to include in the statistics calculation
}}

\item{use_dwell_time}{\itemize{
\item use dwell time instead of calculated polya length for statistics
}}

\item{custom_glm_formula}{\itemize{
\item custom glm formula (when using glm for statistics)
}}
}
\value{
data frame
}
\description{
Calculates polyA statistics for single group of reads (for single transcript)
}
