% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_polya_data.R
\name{drop_polya_list_metadata}
\alias{drop_polya_list_metadata}
\title{Filter metadata from the list of poly(A) predictions}
\usage{
drop_polya_list_metadata(input_list, metadata)
}
\arguments{
\item{input_list}{a list - output of read_polya_multiple() with poly(A) predictions}

\item{metadata}{a character vector with names of metadata columns to keep}
}
\value{
list with filtered metadata
}
\description{
Filter metadata from the list of poly(A) predictions
}
\examples{
\dontrun{

filter_metadata(input_list,metadata=c("sample_name","group"))

}

}
