% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_polya_data.R
\name{filter_polya_list_by_transcript}
\alias{filter_polya_list_by_transcript}
\title{Filter data from the list of poly(A) predictions}
\usage{
filter_polya_list_by_transcript(
  input_list,
  transcript_ids,
  transcript_id_column = "transcript",
  verbose = TRUE
)
}
\arguments{
\item{input_list}{a list - output of read_polya_multiple() with poly(A) predictions}

\item{transcript_id_column}{a character vector with column name with transcript names}

\item{transcript}{a character vector or single character with transcript names to filter}
}
\value{
list with filtered data
}
\description{
Filter data from the list of poly(A) predictions
}
\examples{
\dontrun{

filter_data(input_list,transcript=c("ACTB"),transcript_id_column="transcript")

}
}
