% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_polya_data.R
\name{filter_polya_list_samples}
\alias{filter_polya_list_samples}
\title{Filter samples from the list of poly(A) predictions}
\usage{
filter_polya_list_samples(input_list, samples)
}
\arguments{
\item{input_list}{a list - output of read_polya_multiple() with poly(A) predictions}

\item{samples}{a character vector with sample names to keep}
}
\value{
list with filtered samples
}
\description{
Filter samples from the list of poly(A) predictions
}
\examples{
\dontrun{

filter_polya_list_samples(input_list,samples=c("sample1","sample2"))

}

}
