% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_stats.R
\name{get_nanopolish_processing_info}
\alias{get_nanopolish_processing_info}
\title{Get information about nanopolish processing}
\usage{
get_nanopolish_processing_info(polya_data, grouping_factor = NA)
}
\arguments{
\item{polya_data}{A data.frame or tibble containig unfiltered polya output from Nanopolish,}

\item{grouping_factor}{How to group results (e.g. by sample_name)
best read with \link[nanotail]{read_polya_single} or \link[nanotail]{read_polya_multiple}}
}
\value{
A \link[tibble]{tibble} with counts for each processing state
}
\description{
Process the information returned by \code{nanopolish polya} in the \code{qc_tag} column
}
