% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_polya_data.R
\name{get_references}
\alias{get_references}
\title{Get references from the list of poly(A) predictions}
\usage{
get_references(input_list, reference_column = "reference")
}
\arguments{
\item{input_list}{a list - output of read_polya_multiple() with poly(A) predictions}

\item{reference_column}{a character vector with column name with references}
}
\value{
list with references
}
\description{
Get references from the list of poly(A) predictions
}
\examples{
\dontrun{

get_references(input_list,reference_column="reference")

}
 
}
