% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_polya_data.R
\name{get_transcript_data_from_list}
\alias{get_transcript_data_from_list}
\title{Get data for a specific transcripts from a list of poly(A) predictions}
\usage{
get_transcript_data_from_list(input_list, transcript)
}
\arguments{
\item{input_list}{a list - output of read_polya_multiple() with poly(A) predictions}

\item{transcript}{a character vector or single character with transcript names to filter}
}
\value{
data.frame (tibble) with data for selected transcripts
}
\description{
Get data for a specific transcripts from a list of poly(A) predictions
}
