% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_helper_functions.R
\name{getmode}
\alias{getmode}
\title{Function calculating statistical mode of given vector.}
\usage{
getmode(x, method = "density", na.rm = FALSE)
}
\arguments{
\item{x}{\link{character} data for which the most frequent value is to be calculated
(polya_data column  with the lengths of the poly(A) tails)}

\item{method}{\link{character} "density"/"value"; density mode is computed
by default.}

\item{na.rm}{\link{boolean} parameter defining whether to remove missing values or
not. By a default set to false}
}
\value{
statistical mode of given vector.
}
\description{
Function calculating statistical mode of given vector.
}
\examples{
\dontrun{

getmode(x = polya_data$tail_length,
        method = "density",
        na.rm = FALSE)

}

}
