% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_helper_functions.R
\name{gm_mean}
\alias{gm_mean}
\title{Calculates geometric mean}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{\itemize{
\item input vector
}}

\item{na.rm}{\itemize{
\item should NA values be removed?
}}
}
\description{
Calculates geometric mean
}
