% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_helper_functions.R
\name{is.nanotail_polya_data}
\alias{is.nanotail_polya_data}
\title{Check if provided list is of class nanotail_polya_data}
\usage{
is.nanotail_polya_data(input_list, verbose = T)
}
\arguments{
\item{input_list}{list to be checked}
}
\value{
boolean
}
\description{
Check if provided list is of class nanotail_polya_data
}
\examples{
\dontrun{

is.nanotail_polya_data(input_list)

}

}
