% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_stats.R
\name{kruskal_polya}
\alias{kruskal_polya}
\title{Compute Kruskal-Wallis test on poly(A) data}
\usage{
kruskal_polya(
  input_data,
  grouping_factor = "sample_name",
  transcript_id = "transcript",
  verbose = F,
  verbosity_level = 1
)

kruskal_polya(
  input_data,
  grouping_factor = "sample_name",
  transcript_id = "transcript",
  verbose = F,
  verbosity_level = 1
)
}
\arguments{
\item{input_data}{\itemize{
\item input tibble/data.frame with nanopolish output.
}}

\item{grouping_factor}{\itemize{
\item which column contains group information
}}

\item{transcript_id}{column which transcript ids}

\item{verbose}{verbose output}

\item{verbosity_level}{how verbose the output should be (levels 1 - little verbosity, or 2 - very verbose)}
}
\value{
data.frame with statistis

data.frame with statistis
}
\description{
Compute Kruskal-Wallis test on poly(A) data

Compute Kruskal-Wallis test on poly(A) data
}
\examples{
\dontrun{
polya_table <- nanotail::read_polya_single("nanopolish.tsv")
kruskal_polya(polya_table,grouping_factor="group",transcript_id="transcript",verbose=T)
}
\dontrun{
polya_table <- nanotail::read_polya_single("nanopolish.tsv")
kruskal_polya(polya_table,grouping_factor="group",transcript_id="transcript",verbose=T)
}
}
