% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_helper_functions.R
\name{load_yaml_samples}
\alias{load_yaml_samples}
\title{Load samples information from yaml file}
\usage{
load_yaml_samples(yaml_path)
}
\arguments{
\item{yaml_path}{path to the yaml file}
}
\value{
\link{tibble}
}
\description{
Load samples information from yaml file
}
\examples{
\dontrun{

load_yaml_samples("path/to/yaml/file")

}

}
