% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_helper_functions.R
\docType{data}
\name{nanotail_ggplot2_theme}
\alias{nanotail_ggplot2_theme}
\title{Default theme for ggplot2-based plots in the NanoTail package}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 4.
}
\usage{
nanotail_ggplot2_theme
}
\description{
Default theme for ggplot2-based plots in the NanoTail package
}
\keyword{datasets}
