% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_shiny_app.R
\name{nanoTailApp}
\alias{nanoTailApp}
\title{wrapper for NanoTail Shiny interface}
\usage{
nanoTailApp(polya_table)
}
\arguments{
\item{polya_table}{polyA predictions table. Can be obtained with \link{read_polya_multiple}}
}
\description{
wrapper for NanoTail Shiny interface
}
