% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_stats.R
\name{normalize_counts_to_depth}
\alias{normalize_counts_to_depth}
\title{Normalize counts to sequencingdepth}
\usage{
normalize_counts_to_depth(
  summarized_data,
  raw_data,
  spike_in_data = NULL,
  groupBy,
  force = F
)
}
\arguments{
\item{summarized_data}{\itemize{
\item output of summarize_polya_per_transcript()
}}

\item{raw_data}{\itemize{
\item raw polyA data (loaded with read_polya_single() or read_polya_multiple())
}}

\item{spike_in_data}{\itemize{
\item spike-in data for normalization (optional) (raw polya data loaded with read_polya_single() or read_polya_multiple(), with the  same metadata as raw data)
}}

\item{groupBy}{\itemize{
\item grouping variable
}}

\item{force}{\itemize{
\item force recalculation
}}
}
\value{
data.frame (tibble) with normalized data
}
\description{
Normalize counts to sequencingdepth
}
