% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_annotate.R
\name{parse_gencode_headers}
\alias{parse_gencode_headers}
\title{Parse references provided in the gencode format}
\usage{
parse_gencode_headers(input_vector)
}
\arguments{
\item{input_vector}{vector with gencode-formatted references}
}
\value{
a data.frame with parsed elements:
\itemize{
\item reference: original reference
\item ensembl_transcript_id: ensembl transcript id (without version number)
\item ensembl_gene_id: ensembl gene id (without version number)
\item symbol: gene symbol
}
}
\description{
Parse references provided in the gencode format
}
\examples{
parse_gencode_headers(c("ENST00000456328|ENSG00000223997|OTTHUMG00000000961|OTTHUMT00000002844.2|OTTHUMT00000002844|DDX11L1|202|processed_transcript|","ENST00000450305|ENSG00000223997|OTTHUMG00000000961|OTTHUMT00000002844.2|OTTHUMT00000002844|DDX11L1|202|processed_transcript|"))


}
