% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_plots.R
\name{plot_counts_scatter}
\alias{plot_counts_scatter}
\title{Title}
\usage{
plot_counts_scatter(polya_data_summarized, groupingFactor = NA,
  condition1 = NA, condition2 = NA, ...)
}
\arguments{
\item{polya_data_summarized}{polyA predictions table, summarized using \link{summarize_polya}}

\item{groupingFactor}{name of column used for grouping}

\item{condition1}{first condition to use for plotting}

\item{condition2}{second condition to use for plotting}

\item{...}{parameters passed to .basic_aesthetics function (scale_x_limit_low = NA, scale_x_limit_high = NA, scale_y_limit_low = NA, scale_y_limit_high = NA, color_palette = "Set1",plot_title=NA)
#'}
}
\value{
\link[ggplot2]{ggplot} object
}
\description{
Title
}
