% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_plots.R
\name{plot_polya_distribution}
\alias{plot_polya_distribution}
\title{Title}
\usage{
plot_polya_distribution(polya_data, groupingFactor,
  scale_x_limit_low = NA, scale_x_limit_high = NA,
  color_palette = "Set1", reverse_palette = 0)
}
\arguments{
\item{reverse_palette}{}
}
\description{
Title
}
