% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_plots.R
\name{plot_polya_violin}
\alias{plot_polya_violin}
\title{Plots violin plot of estimated polya lengths, other version}
\usage{
plot_polya_violin(
  polya_data,
  groupingFactor,
  additional_grouping_factor = NA,
  condition1 = NA,
  condition2 = NA,
  violin = FALSE,
  add_points = FALSE,
  max_points = 500,
  add_boxplot = TRUE,
  fill_by = NA,
  auto_scale = T,
  transcript_id,
  ...
)
}
\arguments{
\item{polya_data}{input table with polyA predictions}

\item{groupingFactor}{which factor to use for grouping}

\item{additional_grouping_factor}{additional coloring grouping factor}

\item{condition1}{First condition to include on the plot}

\item{condition2}{Second condition to include on the plot}

\item{violin}{Should violin plot be plotted instead of boxplot?}

\item{add_points}{should individual points be plotted (only if less than max_points). Represented as \link[ggforce]{geom_sina}}

\item{max_points}{maximum number of points to be plotted if add_points is specified}

\item{add_boxplot}{Add boxplot inside violin?}

\item{...}{parameters passed to .basic_aesthetics function (scale_x_limit_low = NA, scale_x_limit_high = NA, scale_y_limit_low = NA, scale_y_limit_high = NA, color_palette = "Set1",plot_title=NA)}
}
\value{
\link[ggplot2]{ggplot} object
}
\description{
Plots violin plot of estimated polya lengths, other version
}
