% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_plots.R
\name{plot_quantiles}
\alias{plot_quantiles}
\title{Plot quantiles}
\usage{
plot_quantiles(
  summarized_data,
  transcript_id,
  transcript_id_column = "transcript",
  groupBy
)
}
\arguments{
\item{summarized_data}{\itemize{
\item summarized data table (output of summarize_polya_per_transcript, with quantiles calculated)
}}

\item{transcript_id}{\itemize{
\item id of transcript to show
}}

\item{transcript_id_column}{\itemize{
\item column with transcript ids
}}

\item{groupBy}{\itemize{
\item which column use for grouping (e.g. with timepoints)
}}
}
\value{
ggplot object
}
\description{
Plot quantiles
}
