% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_plots.R
\name{plot_virtual_gel}
\alias{plot_virtual_gel}
\title{Title}
\usage{
plot_virtual_gel(
  input_data,
  groupingFactor,
  valuesColumn,
  density_bw = 0.6,
  kernel = "gaussian",
  shift_bands = 0,
  kernel_from = 0,
  kernel_to = NA,
  scale_by_size = T,
  scaling_vector = NA
)
}
\arguments{
\item{input_data}{data frame (or tibble) with input values}

\item{groupingFactor}{factor to group input table by}

\item{valuesColumn}{column with numeric values to calculate density on}

\item{density_bw}{bandwidth of density (bw argument to \link{density} function, default=1 )}

\item{kernel}{kenrel to use}

\item{shift_bands}{shift bands by specified value}

\item{kernel_from}{from which value start density estimation}

\item{kernel_to}{to which value perform density estimation}

\item{scale_by_size}{should densities be scaled by sample size}

\item{scaling_vector}{vector containing scaling factors}
}
\value{
list containing plot and density estimates (data.frame)
}
\description{
Title
}
