% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_polya_data.R
\name{polya_list_to_data_frame}
\alias{polya_list_to_data_frame}
\title{Convert a list with poly(A) predictions to long format data.frame with metadata columns (like in the original nanoTail)}
\usage{
polya_list_to_data_frame(input_list)
}
\arguments{
\item{input_list}{a list - output of read_polya_multiple() with poly(A) predictions}
}
\value{
data.frame (tibble)
}
\description{
Convert a list with poly(A) predictions to long format data.frame with metadata columns (like in the original nanoTail)
}
