% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_stats.R
\name{polya_summary_pivot_to_wide}
\alias{polya_summary_pivot_to_wide}
\title{title probably top be removed}
\usage{
polya_summary_pivot_to_wide(polya_data)
}
\arguments{
\item{polya_data}{}
}
\description{
title probably top be removed
}
