% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_polyA_data.R
\name{read_polya_multiple2}
\alias{read_polya_multiple2}
\title{Reads multiple polyA predictions at once}
\usage{
read_polya_multiple2(input_table)
}
\arguments{
\item{input_table}{data.frame or tibble containing samples metadata and paths to files.}
}
\value{
a list containing polyA predictions for all specified samples, with metadata provided in samples_table
}
\description{
Reads multiple polyA predictions at once
}
\examples{
\dontrun{

read_polya_multiple2(example_sample_table)

}

}
