% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_polyA_data.R
\name{read_polya_single2}
\alias{read_polya_single2}
\title{Read Single Nanopolish polyA preditions from file}
\usage{
read_polya_single2(
  polya_path,
  gencode = TRUE,
  sample_name = NA,
  input_type = "nanopolish"
)
}
\arguments{
\item{polya_path}{path to nanopolish output file}

\item{gencode}{are contig names GENCODE-compliant.
Can get transcript names and ensembl_transcript IDs if reads were mapped for example to Gencode reference transcriptome}

\item{sample_name}{sample name (optional), provided as a string.
If specified will be included as an additional column sample_name.}

\item{input_type}{type of input file. Can be one of: \itemize{
\item nanopolish - nanopolish output
\item dorado - dorado output
\item tailfindr - tailfindr output
\item auto - automatically detect input type
}}
}
\value{
a \link[tibble:tibble-package]{tibble} with polya predictions
}
\description{
This is the basic function used to import output from poly(A) prediction programs to R
}
\seealso{
\link{read_polya_multiple}
}
