% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_polyA_data.R
\name{remove_failed_reads}
\alias{remove_failed_reads}
\title{Title}
\usage{
remove_failed_reads(polya_data)
}
\arguments{
\item{polya_data}{}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Title
}
