% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_helper_functions.R
\name{spread_multiple}
\alias{spread_multiple}
\title{Spread multiple columns}
\usage{
spread_multiple(df, key, value)
}
\arguments{
\item{df}{data frame to apply spread on}

\item{key}{as in \link{spread}}

\item{value}{vector of columns to be taken as value for \link{spread}}
}
\value{
\link{tibble}
}
\description{
Spread multiple columns
}
