% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_helper_functions.R
\name{subsample_table}
\alias{subsample_table}
\title{Subsample a date frame}
\usage{
subsample_table(input_table, groupingFactor = NA, subsample = NA)
}
\arguments{
\item{input_table}{input table for subsampling}

\item{groupingFactor}{grouping factor(s)}

\item{subsample}{specify absolute number of rows or fraction to subsample from the data frame (group-wise)}
}
\value{
\link{tibble}
}
\description{
Uses base subsetting and \link{sample} or dplyr \link[dplyr]{sample_n} or \link[dplyr]{sample_frac} to get the subset of the bigger data.frame or tibble
}
