% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_stats.R
\name{summarize_polya}
\alias{summarize_polya}
\title{Title}
\usage{
summarize_polya(polya_data)
}
\arguments{
\item{polya_data}{}
}
\value{
long-format tibble with per-transcript statistics for each sample
}
\description{
Title
}
