% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_stats.R
\name{summarize_polya_per_transcript}
\alias{summarize_polya_per_transcript}
\title{Summarize poly(A) data per transcript}
\usage{
summarize_polya_per_transcript(
  polya_data,
  groupBy = NULL,
  transcript_id_column = transcript,
  summary_functions = list("median", "mean"),
  quantiles = NA
)
}
\arguments{
\item{polya_data}{input table with poly(A) data.}

\item{transcript_id_column}{column with transcript identifier. Default to "transcript"}

\item{summary_functions}{list of summary functions. Set to NA to get only counts per transcript}

\item{quantiles}{vector with quantile values (optional)}

\item{summary_factors}{vector of grouping columns. Set to NULL to omit grouping}
}
\value{
a tibble with summarized poly(A) length data
}
\description{
Summarize poly(A) data per transcript
}
