% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_helper_functions.R
\name{validate_nanotail_polya_data}
\alias{validate_nanotail_polya_data}
\title{Validate if provided object matches requirements of nanotail_polya_data class}
\usage{
validate_nanotail_polya_data(input_list)
}
\arguments{
\item{input_list}{list to be checked}
}
\value{
boolean
}
\description{
Validate if provided object matches requirements of nanotail_polya_data class
}
\examples{
\dontrun{

validate_nanotail_polya_data(input_list)

}

}
