% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_annotate.R
\name{annotate_with_org_packages}
\alias{annotate_with_org_packages}
\title{Title}
\usage{
annotate_with_org_packages(polya_data,
  columns_of_annotation = c("GENENAME", "SYMBOL"),
  keytype = "ENSEMBLTRANS", organism = "mus_musculus")
}
\arguments{
\item{polya_data}{polya data table to annotate}

\item{attributes_to_get}{what annotations should be retrieved. Default = c('external_gene_name','description','transcript_biotype')}

\item{transcript_id}{which column should be matched in the target mart}

\item{mart_to_use}{mart object created with \link[biomaRt]{useMart} or \link[biomaRt]{useEnsembl}}
}
\value{
a \link[tibble]{tibble}
}
\description{
Title
}
