% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_stats.R
\name{calculate_diff_exp_binom}
\alias{calculate_diff_exp_binom}
\title{Performs differential expression analysis}
\usage{
calculate_diff_exp_binom(
  polya_data,
  grouping_factor = NA,
  condition1 = NA,
  condition2 = NA,
  alpha = 0.05,
  summarized_input = FALSE
)
}
\arguments{
\item{polya_data}{polya_data tibble}

\item{grouping_factor}{name of column containing factor with groups for comparison}

\item{condition1}{first condition to compare}

\item{condition2}{second condition to compare}

\item{alpha}{threshold for a pvalue, to treat the result as significant (default = 0.05)}

\item{summarized_input}{is input table already summarized?}
}
\value{
a tibble with differential expression results
}
\description{
Uses counts for each identified transcript to calculate differential expression between specified groups.
This function is a wrapper for \code{\link[edgeR]{binomTest}} from \code{edgeR} package
}
\seealso{
\link[edgeR]{binomTest}
}
