% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_stats.R
\name{calculate_pca}
\alias{calculate_pca}
\title{Calculates PCA using median polya lengths per transcript}
\usage{
calculate_pca(polya_data_summarized, parameter = "polya_median")
}
\arguments{
\item{parameter}{}
}
\description{
Calculates PCA using median polya lengths per transcript
}
