% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_stats.R
\name{calculate_polya_stats}
\alias{calculate_polya_stats}
\title{Calculate basic statistics}
\usage{
calculate_polya_stats(polya_data, min_reads = 0)
}
\arguments{
\item{min_reads}{}
}
\value{
tibble with statistics
}
\description{
Calculate basic statistics
}
