% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_stats.R
\name{calculate_polya_stats}
\alias{calculate_polya_stats}
\title{Calculates basic statistics for polya lengths}
\usage{
calculate_polya_stats(
  polya_data,
  transcript_id_column = "transcript",
  min_reads = 0,
  grouping_factor = "sample_name",
  condition1 = NA,
  condition2 = NA,
  stat_test = "Wilcoxon",
  alpha = 0.05,
  add_summary = TRUE,
  length_summary_to_show = "gm_mean",
  ...
)
}
\arguments{
\item{polya_data}{input table with polyA predictions}

\item{transcript_id_column}{\itemize{
\item name of the column with transcript ids (default = "transcript")
}}

\item{min_reads}{minimum number of reads to include transcript in the analysis}

\item{grouping_factor}{which column defines groups (default: sample_name)}

\item{condition1}{if \code{grouping_factor} has more than 2 levels, which level use for comparison}

\item{condition2}{if \code{grouping_factor} has more than 2 levels, which level use for comparison}

\item{stat_test}{what statistical test to use for testing, currently supports "Wilcoxon" (for \link{wilcox.test}), "KS" (for \link[FSA]{ksTest} from FSA package) or "glm" (for \link{glm})}

\item{alpha}{\itemize{
\item alpha value to consider a hit significant (default - 0.05)
}}

\item{add_summary}{\itemize{
\item add summary (mean polya lengths, counts) to statistics results?
}}

\item{length_summary_to_show}{\itemize{
\item which length summary to show ("median"/"mean"/"gm_mean")
}}

\item{...}{\itemize{
\item additional parameters to pass to .polya_stats (custom_glm_formula,use_dwell_time)
}}
}
\value{
summary table with pvalues and median/mean values associated to each transcript
}
\description{
Takes polyA predictions table as input and checks if there is significant difference in polyA lengths between chosen conditions for each transcript.
By default, Wilcoxon Rank Sum (\link{wilcox.test}) test is used.
}
