% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_stats.R
\name{get_nanopolish_processing_info}
\alias{get_nanopolish_processing_info}
\title{Get information about nanopolish processing}
\usage{
get_nanopolish_processing_info(polya_data, sample = NA)
}
\arguments{
\item{polya_data}{A data.frame or tibble containig unfiltered polya output from Nanopolish,
best read with \code{read_polya_single} or \code{read_polya_multiple}}

\item{sample}{string defining a sample to show, leave empty to show all samples}
}
\value{
A list with elements:
\itemize{
\item number_all_reads - sum of all reads processed
\item number_adapter_reads - number of reads marked as ADAPTER
\item number_noregion_reads - number of reads marked as NOREGION
\item number_number_read_load_failed_reads - number of reads marked as READ_FAILED_LOAD
\item number_number_suffclip_reads - number of reads marked as SUFFCLIP
\item number_number_pass_reads - number of reads marked as PASS
}
}
\description{
Process the information returned by \code{nanopolish polya} in the \code{qc_tag} column
}
