% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_helper_functions.R
\name{gm_mean}
\alias{gm_mean}
\title{Calculates geometric mean}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{input vector}

\item{na.rm}{should NA values be removed?}
}
\value{
geometric mean of values provided as an input
}
\description{
Calculates geometric mean
}
\examples{
a <- rnorm(100,33,5)
gm_mean(a)
}
