% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_shiny_app.R
\name{nanoTailApp}
\alias{nanoTailApp}
\title{Title}
\usage{
nanoTailApp(summary_table, polya_table)
}
\arguments{
\item{summary_table}{\itemize{
\item data frame with summarized polya lengths and statistics per transcript
}}

\item{polya_table}{\itemize{
\item data.table object with polyA data for all reads
}}
}
\description{
Title
}
