% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_plots.R
\name{plot_counts_scatter}
\alias{plot_counts_scatter}
\title{Title}
\usage{
plot_counts_scatter(polya_data_summarized, groupingFactor = NA,
  color_palette = "Set1", reverse_palette = 0, condition1 = NA,
  condition2 = NA, min_counts = 0, max_counts = NA)
}
\arguments{
\item{polya_data_summarized}{polyA predictions table, summarized using \link{summarize_polya}}

\item{groupingFactor}{name of column used for grouping}

\item{color_palette}{RColorBrewer color palette to be used}

\item{reverse_palette}{should the palette be reversed}

\item{condition1}{first condition to use for plotting}

\item{condition2}{second condition to use for plotting}

\item{min_counts}{minimum number of counts to be shown}

\item{max_counts}{maximum number of counts to be shown}
}
\value{
\link[ggplot2]{ggplot} object
}
\description{
Title
}
