% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_plots.R
\name{plot_counts_scatter}
\alias{plot_counts_scatter}
\title{Title}
\usage{
plot_counts_scatter(polya_data_summarized, groupingFactor,
  color_palette = "Set1", reverse_palette = 0, condition1 = NA,
  condition2 = NA, min_counts = 0, max_counts = NA)
}
\arguments{
\item{max_counts}{}
}
\description{
Title
}
