% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_plots.R
\name{plot_MA}
\alias{plot_MA}
\title{Plots MA plot of differential expression analysis}
\usage{
plot_MA(binom_test_output, color_palette = "Set1", reverse_palette = 0)
}
\arguments{
\item{binom_test_output}{a table with output from \link{calculate_diff_exp_binom}}

\item{color_palette}{RColorBrewer palette to use}

\item{reverse_palette}{should the palette be reserved}
}
\value{
\link[ggplot2]{ggplot} object
}
\description{
Plots MA plot of differential expression analysis
}
