% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_plots.R
\name{plot_MA}
\alias{plot_MA}
\title{Plots MA plot of differential expression analysis}
\usage{
plot_MA(input_data, color_palette = "Set1", reverse_palette = FALSE)
}
\arguments{
\item{input_data}{a table with output from \link{calculate_diff_exp_binom}}

\item{color_palette}{RColorBrewer palette to use}

\item{reverse_palette}{should the palette be reserved}
}
\value{
\link[ggplot2]{ggplot} object
}
\description{
Crates simple MA plot, with log10(mean expression) on the X-axis and log2(fold_change) on the Y-axis
}
