% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_plots.R
\name{plot_polya_boxplot}
\alias{plot_polya_boxplot}
\title{Title}
\usage{
plot_polya_boxplot(polya_data, groupingFactor, scale_y_limit_low = NA,
  scale_y_limit_high = NA, color_palette = "Set1",
  reverse_palette = 0, plot_title = "")
}
\arguments{
\item{plot_title}{}
}
\description{
Title
}
