% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_plots.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Plots volcano plot of differential expression analysis}
\usage{
plot_volcano(input_data, transcript_id_column, labels = FALSE,
  nlabels = 10, ...)
}
\arguments{
\item{input_data}{a table with output from \link{calculate_diff_exp_binom} or \link{calculate_polya_stats}}

\item{...}{parameters passed to .basic_aesthetics function (scale_x_limit_low = NA, scale_x_limit_high = NA, scale_y_limit_low = NA, scale_y_limit_high = NA, color_palette = "Set1",plot_title=NA)
#'}
}
\value{
\link[ggplot2]{ggplot} object
}
\description{
Plots volcano plot of differential expression analysis
}
