% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_polya_data.R
\name{polya_list_to_data_frame}
\alias{polya_list_to_data_frame}
\title{Convert a list with poly(A) predictions to long format data.frame with metadata columns (like in the original nanoTail)}
\usage{
polya_list_to_data_frame(input_list)
}
\arguments{
\item{input_list}{a list - output of read_polya_multiple() with poly(A) predictions}
}
\value{
data.frame (tibble)
}
\description{
When data are loaded by the Nanotail package thery are stored as list of class "nanotail_polya_data",
which is convenient for internal processing but not for further analysis, including visualization.
This function converts the nanotail_polya_data list to a data.frame, containing all per-read poly(A) data
with metadata stored as separate columns.
}
\examples{
\dontrun{

polya_list_to_data_frame(input_list)

}

}
