% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_polyA_data.R
\name{read_polya_multiple}
\alias{read_polya_multiple}
\title{Reads multiple samples at once}
\usage{
read_polya_multiple(samples_table, ...)
}
\arguments{
\item{samples_table}{\itemize{
\item data.frame or tibble containing samples description and paths
}}

\item{...}{\itemize{
\item additional parameters to pass to read_polya_single(), like ensembl=\link{TRUE/FALSE}
}}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Reads multiple samples at once
}
