% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_polyA_data.R
\name{read_polya_single}
\alias{read_polya_single}
\title{Read Single Nanopolish polyA predcitions from file}
\usage{
read_polya_single(polya_path, ensembl = TRUE)
}
\arguments{
\item{polya_path}{\itemize{
\item path to Nanopolish output
}}

\item{ensembl}{\itemize{
\item are contig names ENSEMBL-compliant
}}

\item{group_name}{\itemize{
\item group
}}

\item{sample_name}{\itemize{
\item sample name
}}

\item{batch}{\itemize{
\item batch (or other additional variable)
}}
}
\value{
Returns a tibble
}
\description{
Read Single Nanopolish polyA predcitions from file
}
