% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_polyA_data.R
\name{remove_failed_reads}
\alias{remove_failed_reads}
\title{Removes reads which failed during Nanopolish polya processing}
\usage{
remove_failed_reads(polya_data)
}
\arguments{
\item{polya_data}{output table from \link{read_polya_single} or \link{read_polya_multiple}}
}
\value{
a \link[tibble:tibble-package]{tibble} with only reads having qc_tag=='PASS'
}
\description{
Convenient function to quickly remove all reads failing during nanopolish polya processing
}
\seealso{
\link{read_polya_single}, \link{read_polya_multiple}
}
