% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_helper_functions.R
\name{stat_median_line}
\alias{stat_median_line}
\title{Helper function for calculating median stat for violin/boxpolot ggplot plots}
\usage{
stat_median_line(
  mapping = NULL,
  data = NULL,
  geom = "hline",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{}

\item{data}{}

\item{geom}{}

\item{position}{}

\item{na.rm}{}

\item{show.legend}{}

\item{inherit.aes}{}

\item{...}{}
}
\value{

}
\description{
Helper function for calculating median stat for violin/boxpolot ggplot plots
}
