% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polya_stats.R
\name{summarize_polya}
\alias{summarize_polya}
\title{Summarizes input polya table}
\usage{
summarize_polya(
  polya_data,
  summary_factors = c("group"),
  transcript_id_column = c("transcript")
)
}
\arguments{
\item{polya_data}{input table with polyA predictions}

\item{summary_factors}{specifies column used for grouping (default: group)}

\item{transcript_id_column}{specifies which column use as transcript identifier (default: transcript). Set to \code{NULL} to omit per-transcript stats}
}
\value{
long-format \link[tibble]{tibble} with per-transcript statistics for each sample
}
\description{
Summarizes input table with polyA predictions, calculating medians, mean, geometric means and standard deviation values for each transcript (default).
To get overall summary for each sample or group, specify \code{transcript_id_column=NULL}
}
