% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_polya_data.R
\name{summarize_polya_list}
\alias{summarize_polya_list}
\title{Summarize poly(A) predictions from a list}
\usage{
summarize_polya_list(input_list)
}
\arguments{
\item{input_list}{a list - output of read_polya_multiple() with poly(A) predictions}
}
\value{
data.frame (tibble) with summary statistics for each sample
}
\description{
Summarize poly(A) predictions from a list
}
\examples{
\dontrun{

summarize_polya_list(input_list)

}

}
